+++ graphql/resolvers/index.ts
@@ -1,7 +1,8 @@
 export * from "./register"
 export * from "./me"
 export * from "./track"
-export * from "./dashboard"
+export * from "./dashboardPremium"
 export * from "./dashboardMetrics"
+export * from "./dashboardBasic"
 export * from "./uploadTracker"
 export * from "./updateTrack"
